Added Standard Deviation
Standard Deviation is now used as the level of performance of the artifical neural network architectures that are being tested against each other. All architectures will now be printed to the screen and the best one will be re-printed at the end showing their performances as standard deviation. The desired standard deviation is also being printed on each iteration.